+++ changelog.md
@@ -5,6 +5,7 @@
 #### New features
 
 * Added `KymoTrack.plot()` and `KymoTrackGroup.plot()` methods to conveniently plot the coordinates of tracked lines. See the [kymotracking documentation](https://lumicks-pylake.readthedocs.io/en/latest/tutorial/kymotracking.html) for more details.
+* Added `KymoTrackGroup.ensemble_diffusion()` for estimating an average diffusion for a collection of tracks.
 
 ## v0.13.2 | 2022-11-15